auth: Move server code out of astore dir (#961)
Moves auth server implementation, along with its dependencies out of the
astore directory and places it in its own root 'auth' directory. This is
in an attempt to make it a bit more easy to spin up a local astore
server for development testing.

This does not yet change the astore server from including the auth
server. This is just to move the files out, and point imports to the new
location. A followup PR will be created to move the auth server out of
the astore binary.

No code was altered except for import paths and accompanying build file
import path changes as well as `deps` renames.

Tested by building the entire repo (bazel build //...)